make changes to csv file in master
when i was resolving conflict in this file while merging master branch
and feature1 branch, i accidently deleted the changes made in the master
branch as i didn't see the instruction in the homework. I am adding this
change back to the csv file so that it still includes the changes that I
was supposed to keep when merging.